Move PreReconcileCheckResult to separate file for sharing
